FWAUser FWARegisteredUser models
Changed the design of of FWAUser and FWARegisteredUser model. This means changes to the SIteController.php.

I changed a method that was used to construct a FWAUser and FWARegisteredUser. It is now defined as a constructor using php syntax.

The problem that I have with this now concerns the login and register functions.

In the site controller, when the user logs in the FWAUser object is constructed. This causes some processing to occur. This is not neccessary. I was to go back to 'constructing' the object via a function rather than a php constructor I could make the login and register functions static and not do unnecessary processing.